fft_fixes

+++ kornia/filters/filter.py
@@ -318,7 +318,7 @@ def complex_matmul(a: Tensor, b: Tensor, groups: int = 1) -> Tensor:
     imag = a.imag @ b.real + a.real @ b.imag
     real = torch.movedim(real, real.dim() - 1, 2).squeeze(-1)
     imag = torch.movedim(imag, imag.dim() - 1, 2).squeeze(-1)
-    c = torch.zeros(real.shape, dtype=torch.complex64, device=a.device)
+    c = torch.zeros(real.shape, dtype=torch.complex128, device=a.device)
     c.real, c.imag = real, imag
     return c.view(c.size(0), -1, *c.shape[3:])
